fix react context and over loading
